move string extension to new file
